Move diffable datasource code to separate extension
